% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias_design.R
\name{alias_design}
\alias{alias_design}
\title{All Factor Combinations}
\usage{
alias_design(n)
}
\arguments{
\item{n}{A natural number, the number of factors in the experiment}
}
\value{
Returns a matrix with n columns. All entries will be 0, 1, or 2. This will be the "perfect" design matrix for an experiment with these many factors.
}
\description{
This will produce a design matrix for a non-fractional factorial experiment with n number of factors at 3 levels
}
