% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias_dim_calc.R
\name{alias_dim_calc}
\alias{alias_dim_calc}
\title{Number of Generators + 1}
\usage{
alias_dim_calc(x)
}
\arguments{
\item{x}{The number of defining generators}
}
\value{
A single number. This will be the total number of generators in the alias structure built by the defining generators plus 1.
}
\description{
Gives total number of generators for a set number of defining generators plus 1
}
\note{
The reason for the plus one is for simplification in other coding parts of the package
}
