% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias_matrix.R
\name{alias_matrix}
\alias{alias_matrix}
\title{Creates the Alias Matrix}
\usage{
alias_matrix(X)
}
\arguments{
\item{X}{an encoded first order design matrix of 0's, 1's, and 2's with f columns.}
}
\value{
The alias matrix that corresponds to all second order interactions compared to the main effects of X.
}
\description{
Creates the Alias Matrix
}
\examples{
\dontrun{
X <- ran_D(15, 3)
alias_matrix(X)

X1 <- rot(X, c(1,1,1))
alias_matrix(X1)
}
}
