% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_and_counter.R
\name{cat_and_count}
\alias{cat_and_count}
\title{Used to build tabluated incidents}
\usage{
cat_and_count(x, table_values)
}
\arguments{
\item{x}{two column data frame for which the indicents are to be counted}

\item{table_values}{The character values for which the incidents of x are to be counted for}
}
\description{
Used to build tabluated incidents
}
