% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_and_count_partial.R
\name{cat_and_count_partial}
\alias{cat_and_count_partial}
\title{Used to build tabluated incidents allowing for partial rotations}
\usage{
cat_and_count_partial(x, table_values, t2, t3)
}
\arguments{
\item{x}{two column data frame for which the indicents are to be counted}

\item{table_values}{The character values for which the incidents of x are to be counted for}

\item{t2}{Characters for 2/2 level incident tables}

\item{t3}{Characters for 2/3 level incident tables}
}
\description{
Used to build tabluated incidents allowing for partial rotations
}
