% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_copier.R
\name{col_copier}
\alias{col_copier}
\title{Rearrange (and copy) the columns of a matrix}
\usage{
col_copier(d, cols)
}
\arguments{
\item{d}{The design matrix}

\item{cols}{object that can be coerced into a numeric vector such that cols should}
}
\description{
Rearrange (and copy) the columns of a matrix
}
