% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_opt.R
\name{d_opt}
\alias{d_opt}
\title{Determinant of Inverse of Info. Matrix}
\usage{
d_opt(design)
}
\arguments{
\item{design}{The encoded design matrix}
}
\value{
Determant value
}
\description{
Given a deisgn encoded with 0,1,2 this will calculate the value of the inverse information matrix
}
