% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_opt_noinv.R
\name{d_opt_noinv}
\alias{d_opt_noinv}
\title{Determinant of Info. Matrix}
\usage{
d_opt_noinv(design)
}
\arguments{
\item{design}{The encoded design matrix}
}
\value{
Determant value
}
\description{
Given a deisgn encoded with 0,1,2 this will calculate the value of the information matrix
}
\examples{
\dontrun{
j <- ran_D(20, 4)
d_opt_noinv(j)
}
}
