% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoder.R
\name{encoder}
\alias{encoder}
\title{Internal encoding function}
\usage{
encoder(x)
}
\arguments{
\item{x}{a 0, 1, or 2 value}
}
\value{
A 3 element vector that is (1,0,0) for x = 0, (0,1,0) for x = 1, and (0,0,1) for x = 2.
}
\description{
Move along please
}
