% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_incident.R
\name{full_incident}
\alias{full_incident}
\title{Creates List of Frequencies of Occurances}
\usage{
full_incident(e, tt)
}
\arguments{
\item{e}{A matrix where each pair of columns are represented next to each other}

\item{tt}{A string name for all possible comibnations. Needed incase the frequency is 0.}
}
\description{
This is a purely internatalo function that creates tables of incidents for each factor at each level
}
