% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_incident_partial.R
\name{full_incident_partial}
\alias{full_incident_partial}
\title{Creates List of Frequencies of Occurances for Paritial Expansions}
\usage{
full_incident_partial(e, tt1, tt2, tt3)
}
\arguments{
\item{e}{A matrix where each pair of columns are represented next to each other}

\item{tt1}{A string name for all possible comibnations. Needed incase the frequency is 0.}

\item{tt2}{Character strings for 2/2 level factors}

\item{tt3}{Character strings for 3/2 level factors}
}
\description{
This is a purely internatalo function that creates tables of incidents for each factor at each level
}
