% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_matrix_convert.R
\name{g_matrix_convert}
\alias{g_matrix_convert}
\title{Tranforms a 0-1-2 design matrix to a 0-1 matrix}
\usage{
g_matrix_convert(X)
}
\arguments{
\item{X}{The encoded design matrix in 0-1-2's}
}
\value{
Each factor gains two columns, with values (0,0) if the factor is 0, (1,0) if the factor is 1, and (0,1) is the factor is 2.
}
\description{
Tranforms a 0-1-2 design matrix to a 0-1 matrix
}
