% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latin}
\alias{latin}
\title{Naming Effects Tool}
\format{A data frame with 26 rows and 3 variables:
\describe{
  \item{Level 0}{blank}
  \item{Level 1}{single latin letter}
  \item{Level 2}{latin letter followed by a 2}
}}
\usage{
latin
}
\description{
This data.frame gives a tool in which naming effects in a more traditional setting becomes possible inside R.
}
\keyword{datasets}
