% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oa_check.R
\name{oa_check}
\alias{oa_check}
\title{Encodes 2nd order interactions}
\usage{
oa_check(D)
}
\arguments{
\item{D}{an encoded design matrix}
}
\value{
This will return a data frame with the first f columns being that of D and the remaining being the encoded second order interactions. All facotrs are assigned a latin value (A, B, C, etc..).
}
\description{
For a given 3-level design this function creates and encodes all second order interactions
}
