% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_2level_exp.R
\name{opt_2level_exp}
\alias{opt_2level_exp}
\title{Foldover Design to introduce new levels}
\usage{
opt_2level_exp(design, return_n = 5)
}
\arguments{
\item{design}{The two level design coded 0,1 (intercept included) to be expanded}

\item{return_n}{The number of candidate rotation vectors to return. The default is 5.}
}
\value{
This will return a data frame of return_n by f + 1 for f being the number of factors in the original design.
The fist column is 'Det' which is the determinant of the information matrix for the design created using that rotation vector
The remaining columns are the vectors corresponding to the elements of the rotation vector.
}
\description{
Using the foldover design it is possible to increase the number of levels in a design. This function focuses on the optimum strategry to turn a two level design into a three level design.
}
\examples{
d2 <- ran_D2(25, 7)
opt_2level_exp(d2, return_n = 10)
  
}
