% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_partial_2level.R
\name{opt_partial_2level}
\alias{opt_partial_2level}
\title{Foldover Design to introduce new levels to a subset of factors}
\usage{
opt_partial_2level(design, return_n = 5, cl, opt = "Det")
}
\arguments{
\item{design}{The two level design coded 0,1 (intercept included) to be expanded}

\item{return_n}{The number of candidate rotation vectors to return. The default is 5.}

\item{cl}{A vector of the location of columns that should be exapnded from 2 levels to 3 levels.}

\item{opt}{What optimality criterion should be used? Choices are 'Det' for the largest determinants or 'Min_Incident' which returns the max of the min of the incidnets of two factor levels.}
}
\value{
This will return a data frame of return_n by f + 1 for f being the number of factors in the original design.
The fist column is 'Det' which is the determinant of the information matrix for the design created using that rotation vector
The remaining columns are the vectors corresponding to the elements of the rotation vector.
}
\description{
Using the foldover design it is possible to increase the number of levels in a design. This function focuses on the optimum strategry to turn a two level design into a three level design. Or, as in this function, a subset of those 2-level factors to 3-level factors.
}
