% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ran_D.R
\name{ran_D}
\alias{ran_D}
\title{Generate a Random 3-Level Design}
\usage{
ran_D(trt_num, var_num)
}
\arguments{
\item{trt_num}{The maximum number of rows you want your design to have}

\item{var_num}{The number of 3-level factors you would like}
}
\value{
A matrix randomy filled with 0's, 1', and 2's with var_num columns.
}
\description{
This will create a random 3 level design. The design may not be an orthogonal array or even full column rank.
}
\examples{
\donttest{
ran_D(15, 3)
}


}
