% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ran_d2.R
\name{ran_D2}
\alias{ran_D2}
\title{Randomly Build a 2-level Design}
\usage{
ran_D2(n_rows, factors, unique = FALSE)
}
\arguments{
\item{n_rows}{The number of rows you would like in the final design. If unique = TRUE then this is the upper bound.}

\item{factors}{The number of factors you would like in the experiment}

\item{unique}{Should replicate runs be allowed?}
}
\value{
A design data frame with the number of columns being the nubmer of factors.
}
\description{
This fucntion randomly builds a 2-level design that allws for unique or non unique runs.
}
