% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot.R
\name{rot}
\alias{rot}
\title{Generating a New Design Matrix}
\usage{
rot(D, rotation_vector, unique = TRUE)
}
\arguments{
\item{D}{The original design matrix that you would like expanded upon, must come in coded with 0,1,2's}

\item{rotation_vector}{The rotation vector with the same length as the column width of the design matirx. Entries may be 0, 1, or 2.}

\item{unique}{Should rows that are repeated be removed? Default is TRUE}
}
\value{
This will return a data frame such that the triple foldover has been applied to the original design D.
}
\description{
Create the new design matrix from a triple foldover techinque using a rotation vector
}
\examples{
\dontrun{
design <- matrix(c(1,2,1,0,1,1,1,2,2), byrow = TRUE, nrow = 3)
x <- c(1, 2, 0)
rot(design, x)
}
}
