% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submatrix_counter.R
\name{submatrix_counter}
\alias{submatrix_counter}
\title{Check's Alias submatrix for 0's}
\usage{
submatrix_counter(v, alias)
}
\arguments{
\item{v}{The top left of a 2x2 submatrix coordinates}

\item{alias}{An alias matrix of interest}
}
\value{
1 if all four elements are 0, 0 otherwise
}
\description{
This funciton is basically a logic function to see if corrdinates v of design is a submatrix of the alias function that has elements of all 0's
}
\examples{
\dontrun{
d <- ran_D(25, 4)
alias <- alias_matrix(d)
submatrix_counter(c(3,5), alias)
alias_2 <- alias[c(3,4), c(5,6)]
submatrix_counter(c(3,5), alias_2)
}
}
