% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_inc.R
\name{tab_inc}
\alias{tab_inc}
\title{Internal function for tabulating incidents}
\usage{
tab_inc(d)
}
\arguments{
\item{d}{The design to have it's incidnets tabulated}
}
\description{
Generally useful to compare minimum incidents
}
