% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_inc2_partial.R
\name{tab_inc2_partial}
\alias{tab_inc2_partial}
\title{Internal function for tabulating incidents for partial foldovers}
\usage{
tab_inc2_partial(d)
}
\arguments{
\item{d}{The design to have it's incidnets tabulated}
}
\description{
Generally useful to compare minimum incidents
}
