% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tester_d.R
\name{tester_d}
\alias{tester_d}
\title{Design and Determinant Calculator}
\usage{
tester_d(rotv, design)
}
\arguments{
\item{rotv}{The rotation vector with elements 0,1,2 of length f for f being number of factors}

\item{design}{The design matrix to be expanded with elements 0,1,2}
}
\value{
Determinant for the agumented design's information matrix
}
\description{
Finds the determinant for an augmented design's information matrix created within the fucntion
}
