% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tester_d2.R
\name{tester_d2}
\alias{tester_d2}
\title{Alternative for design creation and det calculation}
\usage{
tester_d2(rotv, design, inv = FALSE)
}
\arguments{
\item{rotv}{The rotation vector to be applied column wise.}

\item{design}{The design coded 0,1 for the rotation to be applied to}

\item{inv}{Should the det of the inverse be calculated}
}
\value{
The determinant of the information matrix is returned
}
\description{
This is similar to it's twin tester_d but does not allow one to choose if replications are allowed (they are in this function).
}
