% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tester_inc.R
\name{tester_inc}
\alias{tester_inc}
\title{Wrapper to apply an incident counter}
\usage{
tester_inc(rotv, design)
}
\arguments{
\item{rotv}{The rotation vector to be used}

\item{design}{The original coded 0/1/2 design}
}
\description{
In general this function works similarly to tester_D and tester_D2
}
