% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tester_inc2.R
\name{tester_inc2}
\alias{tester_inc2}
\title{Wrapper to apply an incident counter for 0/1 designs}
\usage{
tester_inc2(rotv, design)
}
\arguments{
\item{rotv}{The rotation vector to be used}

\item{design}{The original coded 0/1/2 design}
}
\description{
In general this function works similarly to tester_D and tester_D2 and is the twin of tester_inc with the exception this is for 0/1 designs only.
}
