% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tester_partial2.R
\name{tester_partial2}
\alias{tester_partial2}
\title{Alternative for design creation and det calculation}
\usage{
tester_partial2(rotv, design, inv = FALSE)
}
\arguments{
\item{rotv}{The rotation vector to be applied column wise.}

\item{design}{The design coded 0,1 for the rotation to be applied to}

\item{inv}{Should the det of the inverse be calculated}
}
\value{
The determinant of the information matrix is returned
}
\description{
This is similar to it's twin tester_d2 but allows for the determinant to be calculated for 2-level and 3-level mixed designs.
}
