% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_rot_vec.R
\name{opt_rot_vec}
\alias{opt_rot_vec}
\title{Finds the optimal foldover plans for 3 level designs}
\usage{
opt_rot_vec(design, return_n = 5)
}
\arguments{
\item{design}{The encoded design matrix using 0,1,2 notation that is to be expanded}

\item{return_n}{How many of the top rotation vectors (with regards to Omega) should be returned? Default is 5.}
}
\description{
For a given design matrix encoded with 0,1,2 this function will return the values of Omega, the determinant of X_1'X_1, and
}
\examples{
\dontrun{
d <- ran_D(25, 4)
opt_rot_vec(d, return_n) 
}

}
