% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tester.R
\name{tester}
\alias{tester}
\title{Design and Omega Calculator}
\usage{
tester(rotv, design)
}
\arguments{
\item{rotv}{The rotation vector with elements 0,1,2 of length f for f being number of factors}

\item{design}{The design matrix to be expanded with elements 0,1,2}
}
\value{
Omega value for the agumented design
}
\description{
Finds the Omega value for an augmented design created within the fucntion
}
